// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] }
// ...
export interface SettingsDocumentDataNavLinksItem {
  /**
   * Label field in *Settings → Navigation Links*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Link Label
   * - **API ID Path**: settings.nav_links[].label
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  label: prismic.KeyTextField

  /**
   * Link field in *Settings → Navigation Links*
   *
   * - **Field Type**: Link
   * - **Placeholder**: Link URL
   * - **API ID Path**: settings.nav_links[].link
   * - **Documentation**: https://prismic.io/docs/fields/link
   */
  link: prismic.LinkField<string, string, unknown, prismic.FieldState, never>
}

/**
 * Item in *Settings → Social Links*
 */
export interface SettingsDocumentDataSocialLinksItem {
  /**
   * Icon field in *Settings → Social Links*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.social_links[].icon
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  icon: prismic.ImageField<never>

  /**
   * Link field in *Settings → Social Links*
   *
   * - **Field Type**: Link
   * - **Placeholder**: Social Link URL
   * - **API ID Path**: settings.social_links[].link
   * - **Documentation**: https://prismic.io/docs/fields/link
   */
  link: prismic.LinkField<string, string, unknown, prismic.FieldState, never>
}

/**
 * Content for Settings documents
 */
interface SettingsDocumentData {
  /**
   * Header Logo field in *Settings*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.header_logo
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  header_logo: prismic.ImageField<never>

  /**
   * Navigation Links field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.nav_links[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  nav_links: prismic.GroupField<Simplify<SettingsDocumentDataNavLinksItem>>

  /**
   * CTA Text field in *Settings*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Request a quote
   * - **API ID Path**: settings.cta_text
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  cta_text: prismic.KeyTextField

  /**
   * CTA Link field in *Settings*
   *
   * - **Field Type**: Link
   * - **Placeholder**: Link URL
   * - **API ID Path**: settings.cta_link
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/link
   */
  cta_link: prismic.LinkField<
    string,
    string,
    unknown,
    prismic.FieldState,
    never
  >

  /**
   * Footer Logo field in *Settings*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.footer_logo
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  footer_logo: prismic.ImageField<never>

  /**
   * Social Links field in *Settings*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: settings.social_links[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  social_links: prismic.GroupField<
    Simplify<SettingsDocumentDataSocialLinksItem>
  >

  /**
   * Contact Email field in *Settings*
   *
   * - **Field Type**: Text
   * - **Placeholder**: info@positivus.com
   * - **API ID Path**: settings.contact_email
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  contact_email: prismic.KeyTextField

  /**
   * Contact Phone field in *Settings*
   *
   * - **Field Type**: Text
   * - **Placeholder**: 555-567-8901
   * - **API ID Path**: settings.contact_phone
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  contact_phone: prismic.KeyTextField

  /**
   * Contact Address field in *Settings*
   *
   * - **Field Type**: Text
   * - **Placeholder**: 1234 Main St...
   * - **API ID Path**: settings.contact_address
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  contact_address: prismic.KeyTextField
  contact_address2: prismic.KeyTextField

  /**
   * Copyright Text field in *Settings*
   *
   * - **Field Type**: Text
   * - **Placeholder**: © 2023 Positivus. All Rights Reserved.
   * - **API ID Path**: settings.copyright_text
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  copyright_text: prismic.KeyTextField
}

/**
 * Settings document from Prismic
 *
 * - **API ID**: `settings`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type SettingsDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithoutUID<
    Simplify<SettingsDocumentData>,
    'settings',
    Lang
  >